pyproject.toml
[tool.poetry]
name = "beastwick18-kitty-background-manager"
version = "0.1.0"
description = ""
authors = ["Brad Culwell <beastwick18@gmail.com>"]
readme = "README.md"

[tool.poetry.scripts]
kittybg = "beastwick18_kitty_background_manager.main:app"

[tool.poetry.dependencies]
python = "^3.10"
typer = {extras = ["all"], version = "^0.4.1"}
Pillow = "^9.1.1"

[tool.poetry.dev-dependencies]
pytest = "^6.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
